// src/app/shared/Model/BCMProgramUpdate.ts
export class bcmprogramModelUpdate {
    bcM_Program_Status: number;
    bcM_Program_Name: string;
    bcM_Program_Description: string;
    bcM_Program_Owner: number;
    bcM_Program_Manager: number;
    bcM_Program_StartDate: Date;
    bcM_Actual_CompletionDate: Date;
    review_Period_Days: number;
    last_Review_Date: Date;
    bcM_Next_Review_Due_Date: Date;
    bcM_Program_Region: string;
    created_By: number;
    created_Date: Date;
    bcM_Program_ID :Number;
}